backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==23.5
Pillow==10.1.0
gunicorn==21.2.0
python-dateutil==2.8.2
